requirements.txt
gpiozero
picamera2
